Move operator to new file to improve versioning
